+++ packages/groq/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @ai-sdk/groq
 
+## 4.0.0-canary.49
+
+### Patch Changes
+
+- 023550e: Deprecate `streamText` result `fullStream` in favor of `stream`.
+
+## 4.0.0-canary.48
+
+### Patch Changes
+
+- Updated dependencies [6c93e36]
+- Updated dependencies [f617ac2]
+  - @ai-sdk/provider-utils@5.0.0-canary.44
+
 ## 4.0.0-canary.47
 
 ### Patch Changes